requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
RPi.GPIO==0.6.5
Werkzeug==0.15.1
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pyserial==3.4
RPi.GPIO==0.6.5
schedule==0.6.0
Werkzeug==0.15.1
